style(network-web-auth): satisfy perfectionist dylint lints
CI's dylint job flagged the new crate against the `perfectionist` lint
library (which clippy does not run). Fix every finding:

- Reorder derive lists to the configured `prefix_then_alphabetical`
  order (`Debug, Default, Clone, ...`).
- Rename the single-letter `R` reporter generic to `Reporter`, bound as
  `self::Reporter`, matching the existing pacquet convention.
- Move the inline `mod tests { ... }` blocks for `capabilities`,
  `generate_qr_code`, and `web_auth_timeout_error` into external
  `tests.rs` files (the crate's `unit_test_file_layout` is `external_only`).
- Add trailing commas to multi-line macro invocations.

No behavior change. Verified clean with the same command CI runs:
`RUSTFLAGS='-D warnings' cargo dylint --all -- --all-targets --workspace`.
