+++ pacquet/crates/network-web-auth/src/capabilities.rs
@@ -252,24 +252,4 @@ fn map_dialoguer_error(error: dialoguer::Error) -> PromptError {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{Clock, Host, StdinIsTty, StdoutIsTty};
-
-    /// The clock reads a real, post-epoch wall-clock value.
-    #[test]
-    fn host_clock_reads_a_non_zero_time() {
-        let now = Host::now_ms();
-        eprintln!("Host::now_ms() = {now}");
-        assert!(now > 0);
-    }
-
-    /// The TTY probes are dispatchable and return a bool. The value
-    /// depends on how the test harness wired stdio, so only its type is
-    /// asserted — the behavioral branches are covered by fakes in the
-    /// `prompt_browser_open` / `with_otp_handling` tests.
-    #[test]
-    fn host_tty_probes_are_callable() {
-        let _: bool = Host::stdin_is_tty();
-        let _: bool = Host::stdout_is_tty();
-    }
-}
+mod tests;
